Fix db migration script

+++ src/migration/1688676100216-addAgentPayg.ts
@@ -7,10 +7,20 @@ export class AddAgentPayg1688676100216 implements MigrationInterface {
   name = 'AddAgentPayg1688676100216';
 
   async up(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(`ALTER TABLE "channel" ADD "agent" character varying NOT NULL DEFAULT ''`);
+    const table = await queryRunner.getTable('channel');
+    const isAgentColumnExist = table.columns.find((column) => column.name === 'agent');
+
+    if (!isAgentColumnExist) {
+      await queryRunner.query(`ALTER TABLE "channel" ADD "agent" character varying NOT NULL DEFAULT ''`);
+    }
   }
 
   async down(queryRunner: QueryRunner): Promise<void> {
-    await queryRunner.query(`ALTER TABLE "channel" DROP COLUMN "agent"`);
+    const table = await queryRunner.getTable('channel');
+    const isAgentColumnExist = table.columns.find((column) => column.name === 'agent');
+
+    if (isAgentColumnExist) {
+      await queryRunner.query(`ALTER TABLE "channel" DROP COLUMN "agent"`);
+    }
   }
 }
